# [PERF] hr_skills{,_slides}: speed up _recompute_completion