--- a/addons/hr_skills_slides/models/slide_channel.py
+++ b/addons/hr_skills_slides/models/slide_channel.py
@@ -12,29 +12,34 @@ class SlideChannelPartner(models.Model):
 
     def _recompute_completion(self):
         res = super(SlideChannelPartner, self)._recompute_completion()
+        completed_membership = self.filtered(lambda m: m.member_status == 'completed')
+        if not completed_membership:
+            return res
         partner_has_completed = {
-            channel_partner.partner_id.id: channel_partner.channel_id for channel_partner in self
-            if channel_partner.member_status == 'completed'}
+            membership.partner_id.id: membership.channel_id
+            for membership in completed_membership
+        }
         employees = self.env['hr.employee'].sudo().search(
-            [('user_id.partner_id', 'in', list(partner_has_completed.keys()))])
+            [('user_id.partner_id', 'in', completed_membership.partner_id.ids)])
 
         if employees:
             HrResumeLine = self.env['hr.resume.line'].sudo()
             line_type = self.env.ref('hr_skills_slides.resume_type_training', raise_if_not_found=False)
             line_type_id = line_type and line_type.id
 
+            lines_for_channel_by_employee = dict(HrResumeLine._read_group([
+                ('employee_id', 'in', employees.ids),
+                ('channel_id', 'in', completed_membership.channel_id.ids),
+                ('line_type_id', '=', line_type_id),
+                ('display_type', '=', 'course')
+            ], ['employee_id'], ['channel_id:array_agg']))
+
+            lines_to_create = []
             for employee in employees:
                 channel = partner_has_completed[employee.user_id.partner_id.id]
 
-                already_added = HrResumeLine.search([
-                    ("employee_id", "in", employees.ids),
-                    ("channel_id", "=", channel.id),
-                    ("line_type_id", "=", line_type_id),
-                    ("display_type", "=", "course")
-                ])
-
-                if not already_added:
-                    HrResumeLine.create({
+                if channel.id not in lines_for_channel_by_employee.get(employee, []):
+                    lines_to_create.append({
                         'employee_id': employee.id,
                         'name': channel.name,
                         'date_start': fields.Date.today(),
@@ -44,6 +49,8 @@ class SlideChannelPartner(models.Model):
                         'display_type': 'course',
                         'channel_id': channel.id
                     })
+            if lines_to_create:
+                HrResumeLine.create(lines_to_create)
         return res
 
     def _send_completed_mail(self):